+++ pulsar-broker/src/main/java/org/apache/pulsar/compaction/CompactorTool.java
@@ -143,7 +143,7 @@ public class CompactorTool {
         MetadataStoreExtended store = MetadataStoreExtended.create(brokerConfig.getMetadataStoreUrl(),
                 MetadataStoreConfig.builder()
                         .sessionTimeoutMillis((int) brokerConfig.getMetadataStoreSessionTimeoutMillis())
-                        .metadataStoreName(MetadataStoreConfig.LOCAL_METADATA_STORE)
+                        .metadataStoreName(MetadataStoreConfig.METADATA_STORE)
                         .build());
 
         @Cleanup
