+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/stats/PrometheusMetricsTest.java
@@ -66,15 +66,12 @@ import org.apache.pulsar.broker.service.persistent.PersistentSubscription;
 import org.apache.pulsar.broker.service.persistent.PersistentTopic;
 import org.apache.pulsar.broker.stats.prometheus.PrometheusMetricsGenerator;
 import org.apache.pulsar.client.api.Consumer;
-import org.apache.pulsar.client.api.Message;
 import org.apache.pulsar.client.api.MessageRoutingMode;
 import org.apache.pulsar.client.api.Producer;
 import org.apache.pulsar.client.api.PulsarClient;
 import org.apache.pulsar.client.api.PulsarClientException;
-import org.apache.pulsar.client.api.Schema;
 import org.apache.pulsar.client.api.SubscriptionType;
 import org.apache.pulsar.compaction.Compactor;
-import org.apache.pulsar.metadata.api.MetadataStoreConfig;
 import org.awaitility.Awaitility;
 import org.mockito.Mockito;
 import org.testng.Assert;
@@ -1523,87 +1520,6 @@ public class PrometheusMetricsTest extends BrokerTestBase {
         });
         consumer1.close();
         consumer2.close();
-    }
-
-
-    @Test
-    public void testMetadataStoreStats() throws Exception {
-        String ns = "prop/ns-abc1";
-        admin.namespaces().createNamespace(ns);
-
-        String topic = "persistent://prop/ns-abc1/metadata-store-" + UUID.randomUUID();
-        String subName = "my-sub1";
-
-        @Cleanup
-        Producer<String> producer = pulsarClient.newProducer(Schema.STRING)
-                .topic(topic).create();
-        @Cleanup
-        Consumer<String> consumer = pulsarClient.newConsumer(Schema.STRING)
-                .topic(topic).subscriptionName(subName).subscribe();
-
-        for (int i = 0; i < 100; i++) {
-            producer.newMessage().value(UUID.randomUUID().toString()).send();
-        }
-
-        for (;;) {
-            Message<String> message = consumer.receive(10, TimeUnit.SECONDS);
-            if (message == null) {
-                break;
-            }
-            consumer.acknowledge(message);
-        }
-
-        ByteArrayOutputStream output = new ByteArrayOutputStream();
-        PrometheusMetricsGenerator.generate(pulsar, false, false, false, false, output);
-        String metricsStr = output.toString();
-        Multimap<String, Metric> metricsMap = parseMetrics(metricsStr);
-
-        Collection<Metric> opsLatency = metricsMap.get("pulsar_metadata_store_ops_latency_ms" + "_sum");
-        Collection<Metric> putBytes = metricsMap.get("pulsar_metadata_store_put_bytes" + "_total");
-
-        Assert.assertTrue(opsLatency.size() > 1);
-        Assert.assertTrue(putBytes.size() > 1);
-
-        for (Metric m : opsLatency) {
-            Assert.assertEquals(m.tags.get("cluster"), "test");
-            String metadataStoreName = m.tags.get("name");
-            Assert.assertNotNull(metadataStoreName);
-            Assert.assertTrue(metadataStoreName.equals(MetadataStoreConfig.LOCAL_METADATA_STORE)
-                    || metadataStoreName.equals(MetadataStoreConfig.CONFIGURATION_METADATA_STORE)
-                    || metadataStoreName.equals(MetadataStoreConfig.STATE_METADATA_STORE));
-            Assert.assertNotNull(m.tags.get("status"));
-
-            if (m.tags.get("status").equals("success")) {
-                if (m.tags.get("type").equals("get")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else if (m.tags.get("type").equals("del")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else if (m.tags.get("type").equals("put")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else {
-                    Assert.fail();
-                }
-            } else {
-                if (m.tags.get("type").equals("get")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else if (m.tags.get("type").equals("del")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else if (m.tags.get("type").equals("put")) {
-                    Assert.assertTrue(m.value >= 0);
-                } else {
-                    Assert.fail();
-                }
-            }
-        }
-        for (Metric m : putBytes) {
-            Assert.assertEquals(m.tags.get("cluster"), "test");
-            String metadataStoreName = m.tags.get("name");
-            Assert.assertNotNull(metadataStoreName);
-            Assert.assertTrue(metadataStoreName.equals(MetadataStoreConfig.LOCAL_METADATA_STORE)
-                    || metadataStoreName.equals(MetadataStoreConfig.CONFIGURATION_METADATA_STORE)
-                    || metadataStoreName.equals(MetadataStoreConfig.STATE_METADATA_STORE));
-            Assert.assertTrue(m.value > 0);
-        }
     }
 
     private void compareCompactionStateCount(List<Metric> cm, double count) {
